requirements.txt
appdata==2.1.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.6
colored==1.4.3
facebook-sdk==3.1.0
idna==3.4
numpy==1.21.6
opencv-python==4.5.5.62
requests==2.28.1
to-requirements.txt==1.3.0
urllib3==1.26.13

pytube~=12.1.2